uefi-raw: IpAddress: remove type duplications; use core::net types
This simplifies things as we have less types to work with. The overall
convenience is improved and people will be less confused with the
automated imports suggested by their IDE.

As the core::Net types are ABI compatible, and it is unlikely that they
will ever change in core, we should go for this simplification.